web/requirements.txt
flask
paho-mqtt
influxdb-client